requirements.txt
flask==3.0.0
werkzeug>=3.0.0
requests==2.26.0
google-generativeai>=0.3.2
python-dotenv
gunicorn==22.0.0
urllib3>=1.26.0
pymongo==4.6.1
dnspython==2.4.2